:sparkles: Refactor useModal
