--- packages/web/src/hooks/useModalState.tsx
@@ -1,24 +1,0 @@
-import { useState } from 'react'
-
-interface Props {
-  onModalOpen?: () => void
-  onModalClose?: () => void
-}
-
-const useModalState = ({
-  onModalOpen = () => {},
-  onModalClose = () => {},
-}: Props = {}) => {
-  const [isModalOpen, setIsModalOpen] = useState(false)
-  const openModal = () => {
-    onModalOpen()
-    setIsModalOpen(true)
-  }
-  const closeModal = () => {
-    onModalClose()
-    setIsModalOpen(false)
-  }
-  return { isModalOpen, openModal, closeModal }
-}
-
-export default useModalState
